delete comment folder and counter comment implementation
