--- comment/pract3.py
@@ -1,43 +1,0 @@
-# Import necessary types for typing
-from typing import List, Tuple, Dict, Optional
-# Import LibCST
-import libcst as cst
-import generate_trees as generator
-
-import libcst.matchers as match
-
-# TODO: Create function definitions related to all necessary user stories (should look somewhat like match_imports)
-# TODO: Modify annotations to have node key as type and key as necessary information
-
-
-def match_Comment(cast_dict):
-    """A function for counting the number of comments statements in a Python program."""
-    Comment_dictionary = {}
-    # Iterate through all of the Python files in a directory
-    for file in cast_dict:
-        # Find CASTs for each of these files
-        cast = cast_dict[file]
-        # Determine number of import statements for each file
-        Comment = match.findall(cast, match.Comment())
-        Comment_dictionary[file] = len(Comment)
-
-    return Comment_dictionary
-
-    for file in cast_dict:
-        count = 0
-        count + Comment_dictionary
-    return count
-
-
-
-
-
-
-if __name__ == "__main__":
-    directory = "/home/mkapfhammer/Documents/Allegheny/2022/Spring/CMPSC481/project-team-1/hello"
-    file_list = generator.find_python_files(directory)
-    string_file_list = generator.read_files(directory, file_list)
-    tree_dict = generator.generate_cast(string_file_list)
-
-    imports_dictionary = match_imports(tree_dict)
-    print(imports_dictionary)
