--- comment/test2.py
@@ -1,3 +1,0 @@
-import pytest
-
-from comment import practice2.py
